backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.16.4
tokenizers==0.15.2
transformers==4.35.2
pypdf2==3.0.1
python-dotenv==1.0.0
openai==1.3.7
pydantic==2.5.0
httpx==0.24.1
